+++ packages/app/src/navigation/SidebarNavigation.vue
@@ -109,7 +109,6 @@ import { useRoute } from 'vue-router'
 import SidebarNavigationHeader from './SidebarNavigationHeader.vue'
 import { useDebounceFn, useWindowSize } from '@vueuse/core'
 import { useLoginConnectStore } from '@packages/frontend-shared/src/store/login-connect-store'
-import { isAllowedFeature } from '@packages/frontend-shared/src/utils/isAllowedFeature'
 
 const { t } = useI18n()
 
@@ -178,9 +177,6 @@ watchEffect(() => {
     return
   }
 
-  const showNewBadge = isAllowedFeature('debugNewBadge', loginConnectStore)
-  const newBadge: Badge = { value: t('sidebar.debug.new'), status: 'success', label: t('sidebar.debug.debugFeature') }
-
   if (props.gql?.currentProject?.cloudProject?.__typename === 'CloudProject'
     && props.gql.currentProject.cloudProject.runByNumber
     && props.online
@@ -188,8 +184,6 @@ watchEffect(() => {
     const { status, totalFailed } = props.gql.currentProject.cloudProject.runByNumber || {}
 
     if (status === 'NOTESTS') {
-      setDebugBadge(showNewBadge ? newBadge : undefined)
-
       return
     }
 
@@ -246,8 +240,6 @@ watchEffect(() => {
 
     return
   }
-
-  setDebugBadge(showNewBadge ? newBadge : undefined)
 })
 
 const navigation = computed<{ name: string, icon: FunctionalComponent, href: string, badge?: Badge }[]>(() => {